refactor(ws4): split ResourceHandlers (1466L) into 4 same-type extensions
Pure move — method bodies byte-identical; only file location changes plus
relaxing the 19 read()-dispatched methods from private->internal (private is
file-scoped, so cross-file dispatch requires internal). Multiset line-diff
confirms zero body changes; golden static-wire snapshot diff = 0.

- ResourceHandlers.swift        : struct shell + ISO8601 ext + read() dispatch
- +CacheEnvelope                : cacheEnvelope/wrapWithCacheEnvelope/encodeExtras
                                  (wrapWithCacheEnvelope manual splicing kept, byte-identical)
- +CatalogRouting  (25 methods) : stock plugins/instruments, session players, workflows
- +StateReaders    (21 methods) : transport/tracks/mixer/project/midi/markers/mcu/health
- +LibraryInventory (5 methods) : disk-cache read + path allowlist + size cap
